Make all formatter classes, and create separate file for blockquote.
