+++ Aztec/Classes/Formatters/BlockquoteFormatter.swift
@@ -1,25 +1,7 @@
 import Foundation
 import UIKit
 
-class Blockquote: NSObject, NSCoding {
-    public func encode(with aCoder: NSCoder) {
-
-    }
-
-    override public init() {
-
-    }
-
-    required public init?(coder aDecoder: NSCoder){
-
-    }
-
-    static func ==(lhs: Blockquote, rhs: Blockquote) -> Bool {
-        return true
-    }
-}
-
-struct BlockquoteFormatter: ParagraphAttributeFormatter {
+class BlockquoteFormatter: ParagraphAttributeFormatter {
 
     let elementType: Libxml2.StandardElementType = .blockquote
